big update

+++ src/main/java/com/app/cfd/services/SuperSetService.java
@@ -2,10 +2,12 @@ package com.app.cfd.services;
 
 import com.app.cfd.daos.SuperSetDao;
 import com.app.cfd.daos.TagDao;
+import com.app.cfd.models.OrderedId;
 import com.app.cfd.models.SuperSet;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.UUID;
 
 @Service
@@ -20,18 +22,26 @@ public class SuperSetService {
 
     public UUID insert(SuperSet superSet) {
         return superSetDao.inTransaction(transactional -> {
-            UUID superSetId = transactional.insert(superSet.getName(),
-                    superSet.getDescription(), superSet.getReps());
-            insertSetSuperSetJoins(superSetId, superSet.getSetIds());
+            UUID superSetId = transactional.insert(superSet);
+            if (superSet.getOrderedSetIds() != null) {
+                insertSetSuperSetJoins(superSetId, superSet.getOrderedSetIds());
+            }
             return superSetId;
         });
     }
 
-    public SuperSet getSuperSet(UUID id) {
-        SuperSet superSet = superSetDao.getSuperSet(id);
-        List<UUID> setIds = superSetDao.getSetIdsBySuperSetId(id);
-        superSet.setSetIds(setIds);
+    public SuperSet findById(UUID id) {
+        SuperSet superSet = superSetDao.findById(id);
+        List<OrderedId> orderedSetIds = superSetDao.getOrderedSetIdsBySuperSetId(id);
+        superSet.setOrderedSetIds(orderedSetIds);
         return superSet;
+    }
+
+    public List<SuperSet> getAllSuperSets() {
+        return superSetDao.getAllSuperSets().stream().peek(superSet -> {
+            List<OrderedId> orderedSetIds = superSetDao.getOrderedSetIdsBySuperSetId(superSet.getId());
+            superSet.setOrderedSetIds(orderedSetIds);
+        }).toList();
     }
 
     public void deleteById(UUID id) {
@@ -44,17 +54,24 @@ public class SuperSetService {
     }
 
     public void updateSuperSet(SuperSet superSet) {
-        superSetDao.useTransaction(transactional -> {
-            List<UUID> oldSetIds = transactional.getSetIdsBySuperSetId(superSet.getId());
-            transactional.update(superSet.getId(), superSet.getName(), superSet.getDescription(), superSet.getReps());
-            deleteSetSuperSetJoins(superSet.getId(), oldSetIds);
-            insertSetSuperSetJoins(superSet.getId(), superSet.getSetIds());
-        });
+        if (Objects.equals(superSet.getUserId(), superSetDao.findById(superSet.getId()).getUserId())) {
+            superSetDao.useTransaction(transactional -> {
+                List<UUID> oldSetIds = transactional.getSetIdsBySuperSetId(superSet.getId());
+                transactional.update(superSet);
+                deleteSetSuperSetJoins(superSet.getId(), oldSetIds);
+                if (superSet.getOrderedSetIds() != null) {
+                    insertSetSuperSetJoins(superSet.getId(), superSet.getOrderedSetIds());
+                }
+            });
+        } else {
+            insert(superSet);
+        }
+
     }
 
-    private void insertSetSuperSetJoins(UUID superSetId, List<UUID> setIds) {
-        for (UUID setId : setIds) {
-            superSetDao.insertSetSuperSetLink(setId, superSetId);
+    private void insertSetSuperSetJoins(UUID superSetId, List<OrderedId> orderedSetIds) {
+        for (OrderedId orderedId : orderedSetIds) {
+            superSetDao.insertSetSuperSetLink(orderedId.getId(), orderedId.getOrder(), superSetId);
         }
     }
